// apps/web/src/shared/api/generated/model/bookReadingControllerGetReadingHistoryParams.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { BookReadingControllerGetReadingHistoryActivityRange } from "./bookReadingControllerGetReadingHistoryActivityRange";
import type { BookReadingControllerGetReadingHistorySort } from "./bookReadingControllerGetReadingHistorySort";

export type BookReadingControllerGetReadingHistoryParams = {
  activityRange?: BookReadingControllerGetReadingHistoryActivityRange;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @minimum 1
   * @maximum 9007199254740991
   */
  page?: number;
  sort?: BookReadingControllerGetReadingHistorySort;
};